edje - reduce memory footprint by rearranging structs and types
this should cut some memory used by edje by using smaller types like
shorts instead of ints where we just dont need a full int range and
short will do, and re-ordering in memory data soit packs better when
accoutning for alignment

+++ src/lib/edje/edje_calc.c
@@ -868,7 +868,7 @@ _edje_recalc(Edje *ed)
 void
 _edje_recalc_do(Edje *ed)
 {
-   unsigned int i;
+   unsigned short i;
    Eina_Bool need_calc;
 
 // XXX: dont need this with current smart calc infra. remove me later
